requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
folium>=0.14.0
tqdm>=4.65.0
pillow>=9.5.0
